refactor import registration to own method
